// src/app/model/student-detail-report.ts
export interface StudentDetailReport {
  name: string;
  className: string;
  classId: number;
  status: string;
  joinedDate: string;
  studentId: number;
  firstName: string;
  lastName: string;
  gender: string;
  motherName: string;
  fatherName: string;
  motherPhoneNumber: string;
  fatherPhoneNumber: string;
  motherOccupation: string;
  fatherOccupation: string;
  doorNumber: string;
  street: string;
  addressLine: string;
  city: string;
  state: string;
  pinCode: string;
  assignmentsAssigned: number;
  assignmentsCompleted: number;
  assignmentsPassed: number;
}